+++ Frends.Files.Read/Frends.Files.Read/Definitions/Input.cs
@@ -8,7 +8,7 @@ namespace Frends.Files.Read.Definitions;
 public class Input
 {
     /// <summary>
-    /// Full path of the target file to be written
+    /// Full path of the target file to be read.
     /// </summary>
     /// <example>c:\temp\foo.txt</example>
     [DisplayFormat(DataFormatString = "Text")]