+++ Frends.Files.Read/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [1.1.0] - 2024-10-14
+### Fixed
+- Fixed documentation for Path parameter.
+### Added
+- Added Windows1252 encoding to the FileEncoding enum.
+- Added different sizes to the Task result.
+
 ## [1.0.0] - 2023-02-28
 ### Added
 - Initial implementation